Fix: separar lógica de cancelación en CancelContent y envolver con Suspense
